gst1-plugins-good: Apply 'matroskademux: Start stream time at zero' only on wpe-2.22
This patch was very old (2017) and was causing problems on progressive video
seeks in wpe-2.28, wpe-2.38 and wpe-2.42. Removing it doesn't affect the
original "YouTube TV 2018 20. VideoBufferSize" test that it indended to fix
in the first place (the test passes anyway).

The patch is kept in wpe-2.22 because the progressive video seek issue happens
in that version with and without the patch, and I couldn't test if removing the
patch breaks test 20 (I would need to customize wpe to report support for vp9
in order to pass the tests). As wpe-2.22 is a legacy version, I'm not investing
any time to debug the issue there.

See: https://github.com/WebPlatformForEmbedded/WPEWebKit/issues/1268